added todo function list

+++ Cudheart/Cudheart/Arrays/VectorOps.cuh
@@ -27,7 +27,7 @@ namespace Cudheart::VectorOps {
 	// if you do, prepare for trouble and make it double
 	template <typename T>
 	Vector<T>* arange(T start, T end, T jump) {
-		int len = ((end - start) / jump);
+		int len = (int)((end - start) / jump);
 		Vector<T>* out = empty<T>(len);
 
 		for (int i = 0; start < end; start += jump) {
